edit field type(int=>int64)

+++ app/models/comment.go
@@ -5,10 +5,10 @@ import (
 )
 
 type Comment struct {
-	Id        int
+	Id        int64
 	Body      string
 	Commenter string
-	PostId    int
+	PostId    int // Foreign key for Post (belongs to)
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
